fixed issues from missing_entries.csv
ID 339 and 340 - removed duplicate results keeping the result with the most cases.
ID6 - fixed the HR and CI which were giving a negative SE. This should be fine now. Cells Y512 to AD512
ID202 - this is 13109 sets of twins so 26218 for the combined analysis, 12312 for the men and 13906 for the women. See table 1 in the paper https://academic.oup.com/aje/article/166/3/255/117073